django_theme/requirements.txt
absl-py==2.0.0
amqp==2.6.1
asgiref==3.3.4
asttokens==2.4.1
astunparse==1.6.3
billiard==3.6.4.0
cachetools==5.3.2
celery==4.4.7
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
Django==3.2
django-crispy-forms==1.11.2
exceptiongroup==1.1.3
executing==2.0.1
flatbuffers==23.5.26
fonttools==4.44.0
gast==0.4.0
gdcm==1.1
gevent==23.9.1
google-auth==2.23.4
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.0.1
grpcio==1.59.2
h5py==3.10.0
idna==3.4
ipython==8.17.2
jedi==0.19.1
joblib==1.3.2
keract==4.5.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
kombu==4.6.11
libclang==16.0.6
llvmlite==0.41.1
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
numba==0.58.1
numpy==1.26.1
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.3
parso==0.8.3
Pillow==10.1.0
prompt-toolkit==3.0.40
protobuf==3.19.6
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydicom==2.3.0
Pygments==2.16.1
pylibjpeg==1.4.0
pylibjpeg-libjpeg==1.3.4
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
selenium==3.141.0
shap==0.42.0
SimpleITK==2.3.1
six==1.16.0
sklearn==0.0
slicer==0.0.7
sqlparse==0.4.1
stack-data==0.6.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
tqdm==4.66.1
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.4
vine==1.3.0
wcwidth==0.2.9
Werkzeug==3.0.1
wrapt==1.16.0
zope.event==5.0
zope.interface==6.1